[CMA-98] fix : 입력된 시간 적용 안되는 버그 fix
